+++ engine/h2shared/r_light.c
@@ -78,7 +78,7 @@ DYNAMIC LIGHTS
 R_MarkLights
 =============
 */
-void R_MarkLights (dlight_t *light, int bit, mnode_t *node)
+void R_MarkLights (dlight_t *light, unsigned long long bit, mnode_t *node)
 {
 	mplane_t	*splitplane;
 	float		dist;
@@ -156,7 +156,7 @@ void R_PushDlights (void)
 	{
 		if (l->die < cl.time || !l->radius)
 			continue;
-		R_MarkLights ( l, 1<<i, cl.worldmodel->nodes );
+		R_MarkLights ( l, 1ULL<<i, cl.worldmodel->nodes );
 	}
 }
