+++ engine/hexen2/render.h
@@ -95,7 +95,7 @@ typedef struct entity_s
 						// found in an active leaf
 
 	int			dlightframe;	// dynamic lighting
-	int			dlightbits;
+	unsigned long long			dlightbits;	/* one bit per dlight; MAX_DLIGHTS <= 64. uhexen2-liqz */
 
 // FIXME: could turn these into a union
 	int			trivial_accept;
